Add CLDF versions of datasets
The metadata file is based on the standard Wordlist-metadata.json from cldf

The conversion added form IDs and made Cognateset_IDs global, in addition
to the obvious (and actually unnecessary, it could be done by specifying
the alternatives in the metadata file) steps of csv→tsv conversion and
renaming column headers. It was done using the following code:

for file in *.tsv
do
  x=`basename $file .tsv`
  rm -rf $x
  mkdir $x
  awk 'BEGIN {FS="\t"; OFS=","; print "ID,Cognateset_ID,Language_ID,ISO639P3code,Gloss,Concepticon_ID,Parameter_ID,Form,Class,Segments,Comment"}
    FNR>1 { print FNR-1 , $5"-"$7 , $1 , $2 , "\"" $3  "\"" , $4 , $5 , $6 , $7 , $8 , $9 }' $file > $x/forms.csv
  cp Wordlist-metadata.json $x/
done

inside the datasets/ directory.